// ClientApp/src/app/models/baseline-station-shortage-comment.ts
export class BaselineStationShortageComment {
    ME_PART_NUMBER: string;
    STATION: string;
    KEYWORD_DESCRIPTION: string;
    ON_HAND: number;
    REPL_STA: string;
    REPL_STA_ONHAND: number
    DASH8: string;
    MIN_DATE: Date;
    NOSE: string;
    REQUIRED_IND: string;
    IAG_SCHD: string;
    MXP_THRESHOLD: number;
    DaysInQueue: string;
    COMM: string;
    CHKKEY: string;
}